refactor: clean up actors and consolidate browser runtime
