--- api/apps/steel-browser/api/src/browser-runtime/types.ts
@@ -1,1 +1,0 @@
-export * from "./drivers/types.js";
