--- api/src/browser-runtime/__tests__/adapter.comprehensive.test.ts
@@ -1,201 +1,0 @@
-import { describe, it, expect, beforeEach, vi, afterEach } from "vitest";
-import { XStateAdapter } from "../adapter.js";
-import { BrowserRuntime } from "../facade/browser-runtime.js";
-import { MockLauncher } from "../drivers/mock-launcher.js";
-import { FastifyBaseLogger } from "fastify";
-import { pino } from "pino";
-import { BrowserLauncherOptions } from "../../types/browser.js";
-import { isSimilarConfig } from "../../services/cdp/utils/validation.js";
-
-// Mock validation utils
-vi.mock("../../services/cdp/utils/validation.js", () => ({
-  isSimilarConfig: vi.fn().mockResolvedValue(true),
-}));
-
-describe("XStateAdapter Comprehensive", () => {
-  let mockLogger: FastifyBaseLogger;
-  let launcher: MockLauncher;
-  let runtime: BrowserRuntime;
-  let adapter: XStateAdapter;
-  let mockInstrumentationLogger: any;
-
-  beforeEach(() => {
-    mockLogger = pino({ level: "silent" }) as any;
-    mockInstrumentationLogger = {
-      on: vi.fn(),
-      record: vi.fn(),
-      resetContext: vi.fn(),
-    };
-
-    launcher = new MockLauncher();
-    runtime = new BrowserRuntime({
-      launcher,
-      appLogger: mockLogger,
-      instrumentationLogger: mockInstrumentationLogger,
-    });
-
-    adapter = new XStateAdapter(runtime, mockLogger, mockInstrumentationLogger, {
-      keepAlive: false,
-    });
-  });
-
-  afterEach(async () => {
-    await adapter.shutdown().catch(() => {});
-  });
-
-  describe("Concurrency Protection", () => {
-    it("should serialize concurrent launch calls and return existing browser", async () => {
-      // Spy on start
-      const startSpy = vi.spyOn(runtime, "start");
-
-      // First launch
-      const launch1 = adapter.launch({ options: { headless: true } } as any);
-      // Second launch while first is in progress
-      const launch2 = adapter.launch({ options: { headless: true } } as any);
-
-      const [browser1, browser2] = await Promise.all([launch1, launch2]);
-
-      expect(browser1).toBeDefined();
-      expect(browser2).toBe(browser1);
-
-      // Should only call runtime.start once because of mutex and existing browser check
-      expect(startSpy).toHaveBeenCalledTimes(1);
-    });
-
-    it("should serialize startNewSession calls", async () => {
-      const stopSpy = vi.spyOn(runtime, "stop");
-      const startSpy = vi.spyOn(runtime, "start");
-
-      // Start two sessions concurrently
-      const session1 = adapter.startNewSession({ options: { sessionId: "s1" } } as any);
-      const session2 = adapter.startNewSession({ options: { sessionId: "s2" } } as any);
-
-      await Promise.all([session1, session2]);
-
-      // First session: stop (previous or initial), start
-      // Second session: stop (previous), start
-      expect(startSpy).toHaveBeenCalledTimes(2);
-      expect(stopSpy).toHaveBeenCalledTimes(2);
-    });
-
-    it("should prevent race condition between endSession and startNewSession", async () => {
-      await adapter.launch({ options: { headless: true } } as any);
-
-      const endPromise = adapter.endSession();
-      const startPromise = adapter.startNewSession({ options: { sessionId: "new" } } as any);
-
-      await Promise.all([endPromise, startPromise]);
-
-      expect(adapter.isRunning()).toBe(true);
-    });
-  });
-
-  describe("Config Similarity", () => {
-    it("should reuse browser when config is similar", async () => {
-      vi.mocked(isSimilarConfig).mockResolvedValue(true);
-      const startSpy = vi.spyOn(runtime, "start");
-
-      const config1 = { options: { headless: true }, userDataDir: "/tmp/1" };
-      const config2 = { options: { headless: true }, userDataDir: "/tmp/1" };
-
-      await adapter.launch(config1 as any);
-      await adapter.launch(config2 as any);
-
-      expect(startSpy).toHaveBeenCalledTimes(1);
-    });
-
-    it("should restart session when config differs", async () => {
-      vi.mocked(isSimilarConfig).mockResolvedValue(false);
-      const stopSpy = vi.spyOn(runtime, "stop");
-      const startSpy = vi.spyOn(runtime, "start");
-
-      const config1 = { options: { headless: true }, userDataDir: "/tmp/1" };
-      const config2 = { options: { headless: false }, userDataDir: "/tmp/2" };
-
-      await adapter.launch(config1 as any);
-      await adapter.launch(config2 as any);
-
-      expect(stopSpy).toHaveBeenCalledTimes(1);
-      expect(startSpy).toHaveBeenCalledTimes(2);
-    });
-  });
-
-  describe("KeepAlive Behavior", () => {
-    it("should relaunch after endSession when keepAlive is true", async () => {
-      const keepAliveAdapter = new XStateAdapter(runtime, mockLogger, mockInstrumentationLogger, {
-        keepAlive: true,
-        defaultLaunchConfig: { options: { sessionId: "default" } } as any,
-      });
-
-      const startSpy = vi.spyOn(runtime, "start");
-
-      await keepAliveAdapter.launch({ options: { sessionId: "initial" } } as any);
-      expect(startSpy).toHaveBeenCalledTimes(1);
-
-      await keepAliveAdapter.endSession();
-
-      // Should have relaunched with default config
-      expect(startSpy).toHaveBeenCalledTimes(2);
-      expect(keepAliveAdapter.isRunning()).toBe(true);
-    });
-
-    it("should NOT relaunch after endSession when keepAlive is false", async () => {
-      const startSpy = vi.spyOn(runtime, "start");
-
-      await adapter.launch({ options: { sessionId: "initial" } } as any);
-      expect(startSpy).toHaveBeenCalledTimes(1);
-
-      await adapter.endSession();
-
-      expect(startSpy).toHaveBeenCalledTimes(1);
-      expect(adapter.isRunning()).toBe(false);
-    });
-  });
-
-  describe("Plugin Lifecycle", () => {
-    it("should call plugin lifecycle methods", async () => {
-      const mockPlugin = {
-        name: "test-plugin",
-        onBrowserLaunch: vi.fn(),
-        onBrowserReady: vi.fn(),
-        onSessionEnd: vi.fn(),
-        onBrowserClose: vi.fn(),
-        onShutdown: vi.fn(),
-      };
-
-      adapter.registerPlugin(mockPlugin as any);
-
-      await adapter.launch({ options: { headless: true } } as any);
-
-      // Wait for async ready hooks
-      await new Promise((resolve) => setTimeout(resolve, 50));
-
-      expect(mockPlugin.onBrowserLaunch).toHaveBeenCalled();
-      expect(mockPlugin.onBrowserReady).toHaveBeenCalled();
-
-      await adapter.endSession();
-      expect(mockPlugin.onSessionEnd).toHaveBeenCalled();
-      expect(mockPlugin.onBrowserClose).toHaveBeenCalled();
-
-      await adapter.shutdown();
-      expect(mockPlugin.onShutdown).toHaveBeenCalled();
-    });
-  });
-
-  describe("Browser State Extraction", () => {
-    it("should merge storage data from multiple pages", async () => {
-      await adapter.launch({ options: { headless: true }, userDataDir: "/tmp/data" } as any);
-
-      const browser = adapter.getBrowserInstance()!;
-      const page1 = await browser.newPage();
-      const page2 = await browser.newPage();
-
-      // In a real test we'd mock extractStorageForPage, but XStateAdapter
-      // uses the actual implementation which we can't easily mock here
-      // without more setup. Let's just verify it calls the method.
-      const state = await adapter.getBrowserState();
-      expect(state).toBeDefined();
-      expect(state.cookies).toBeDefined();
-    });
-  });
-});
